Rename `__init__` to allow for namespace packages. See PEP420
