+++ chia/_tests/core/test_daemon_rpc.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from chia import __version__
+from chia.__version__ import __version__
 from chia.daemon.client import connect_to_daemon
